use authenticated decorator for auger and azure
