+++ a2ml/api/auger/dataset.py
@@ -1,5 +1,6 @@
 from .impl.dataset import DataSet
-from .impl.decorators import error_handler, authenticated, with_project
+from .impl.decorators import with_project
+from a2ml.api.utils.decorators import error_handler, authenticated
 from .impl.cloud.rest_api import RestApi
 from .credentials import Credentials
 from .config import AugerConfig
@@ -8,9 +9,9 @@ from .config import AugerConfig
 class AugerDataset(object):
     def __init__(self, ctx):
         self.ctx = ctx
-        credentials = Credentials(ctx).load()
+        self.credentials = Credentials(ctx).load()
         self.ctx.rest_api = RestApi(
-            credentials.api_url, credentials.token)
+            self.credentials.api_url, self.credentials.token)
 
     @error_handler
     @authenticated
