+++ a2ml/api/auger/experiment.py
@@ -1,5 +1,6 @@
 from .impl.cloud.rest_api import RestApi
-from .impl.decorators import error_handler, authenticated, with_dataset
+from .impl.decorators import with_dataset
+from a2ml.api.utils.decorators import error_handler, authenticated
 from .impl.experiment import Experiment
 from .impl.exceptions import AugerException
 from a2ml.api.utils.formatter import print_table
@@ -10,9 +11,9 @@ class AugerExperiment(object):
 
     def __init__(self, ctx):
         self.ctx = ctx
-        credentials = Credentials(ctx).load()
+        self.credentials = Credentials(ctx).load()
         self.ctx.rest_api = RestApi(
-            credentials.api_url, credentials.token)
+            self.credentials.api_url, self.credentials.token)
 
     @error_handler
     @authenticated
