+++ a2ml/api/auger/impl/exceptions.py
@@ -1,3 +1,1 @@
 class AugerException(Exception): pass
-
-class NotAuthenticatedException(AugerException): pass
