+++ a2ml/api/azure/experiment.py
@@ -12,9 +12,11 @@ from azureml.core.compute import ComputeTarget
 from azureml.train.automl import AutoMLConfig
 from azureml.train.automl.run import AutoMLRun
 from azureml.automl.core.featurization import FeaturizationConfig
+
 from .project import AzureProject
 from .exceptions import AzureException
-from .decorators import error_handler
+from a2ml.api.utils.decorators import error_handler, authenticated
+from .credentials import Credentials
 from a2ml.api.utils.formatter import print_table
 from a2ml.api.azure.dataset import AzureDataset
 from a2ml.api.utils import fsclient
@@ -25,8 +27,10 @@ class AzureExperiment(object):
     def __init__(self, ctx):
         super(AzureExperiment, self).__init__()
         self.ctx = ctx
+        self.credentials = Credentials(self.ctx).load()
 
-    @error_handler    
+    @error_handler
+    @authenticated
     def list(self):
         ws = AzureProject(self.ctx)._get_ws()
         experiments = Experiment.list(workspace=ws)
@@ -37,7 +41,8 @@ class AzureExperiment(object):
         self.ctx.log('%s Experiment(s) listed' % str(nexperiments))
         return {'experiments': experiments}
 
-    @error_handler    
+    @error_handler
+    @authenticated    
     def start(self):
         dataset_name = self.ctx.config.get('dataset', None)
         if dataset_name is None:
@@ -127,7 +132,8 @@ class AzureExperiment(object):
 
         return {'experiment_name': experiment_name, 'run_id': run.run_id}
 
-    @error_handler    
+    @error_handler
+    @authenticated    
     def stop(self, run_id = None):
         ws = AzureProject(self.ctx)._get_ws()
         experiment_name = self.ctx.config.get('experiment/name', None)
@@ -143,7 +149,8 @@ class AzureExperiment(object):
         run.cancel()
         return {'stopped': experiment_name}
 
-    @error_handler    
+    @error_handler
+    @authenticated    
     def leaderboard(self, run_id = None):
         ws = AzureProject(self.ctx)._get_ws()
         experiment_name = self.ctx.config.get('experiment/name', None)
@@ -173,11 +180,13 @@ class AzureExperiment(object):
             'leaderboard': leaderboard,
             'status': status }
 
-    @error_handler        
+    @error_handler 
+    @authenticated       
     def get_experiment_settings(self):
         return 
                 
-    @error_handler        
+    @error_handler 
+    @authenticated       
     def history(self):
         ws = AzureProject(self.ctx)._get_ws()
         experiment_name = self.ctx.config.get('experiment/name', None)
