+++ a2ml/api/azure/model.py
@@ -7,19 +7,23 @@ from azureml.core.webservice import Webservice
 from azureml.core.webservice import AciWebservice
 from azureml.exceptions import WebserviceException
 from azureml.train.automl.run import AutoMLRun
+
 from .project import AzureProject
 from .exceptions import AzureException
-from .decorators import error_handler
 from a2ml.api.utils.dataframe import DataFrame
 from a2ml.api.utils import fsclient
+from a2ml.api.utils.decorators import error_handler, authenticated
+from .credentials import Credentials
 
 class AzureModel(object):
 
     def __init__(self, ctx):
         super(AzureModel, self).__init__()
         self.ctx = ctx
+        self.credentials = Credentials(self.ctx).load()
 
     @error_handler
+    @authenticated
     def deploy(self, model_id, locally):
         ws = AzureProject(self.ctx)._get_ws()
         experiment_name = self.ctx.config.get('experiment/name', None)
@@ -77,6 +81,7 @@ class AzureModel(object):
         return {'model_id': model_id, 'aci_service_name': aci_service_name}
 
     @error_handler
+    @authenticated
     def predict(self, filename, model_id, threshold, locally, data, columns):
         ws = AzureProject(self.ctx)._get_ws()
         experiment_name = self.ctx.config.get('experiment/name', None)
@@ -116,6 +121,7 @@ class AzureModel(object):
         return {'predicted': predicted}
 
     @error_handler
+    @authenticated
     def actual(self, filename, model_id):
         pass
